requirements.txt
aiohttp==3.8.1
discord.py
bs4
requests
